모든 length 속성에 .toInt() 완전 적용
- 이미지 크기 계산에서도 .toInt() 적용
- List<int>와 Uint8List의 length 속성이 num을 반환하는 문제 해결
- 모든 파일 크기 계산에서 타입 안전성 완전 확보
- 자동 단가 계산 기능이 모든 파일 타입에서 안정적으로 작동

+++ lib/features/post_system/screens/post_place_screen.dart
@@ -178,9 +178,9 @@ class _PostPlaceScreenState extends State<PostPlaceScreen> {
     // 이미지 크기 계산
     for (var image in _selectedImages) {
       if (image is List<int>) {
-        totalSize += image.length;
+        totalSize += image.length.toInt();
       } else if (image is Uint8List) {
-        totalSize += image.length;
+        totalSize += image.length.toInt();
       }
     }
